Extract iframe-check core
